adds new lock screen widget
